Move renderTree out of the module into SchemaNodeSelector.tsx
